added Beer methods
