+++ models/schema.sql
@@ -1,11 +1,7 @@
+DROP FOREIGN KEY 
+
 DROP DATABASE IF EXISTS beergenie_db;
 CREATE DATABASE beergenie_db;
 
-USE beergenie_db;
-
-CREATE TABLE favorites(
-beer_id 
-user_id AUTO_INCREMENT
-)
