+++ public/js/search.js
@@ -10,4 +10,4 @@ function renderFavorites ({beer}) {
 
     
 
-}
+};
