+++ public/js/welcome.js
@@ -6,7 +6,8 @@
 
 
 $(document).on('click','button.search','/search.html' );
-$(document).on('submit', '#form-search','/search.html' )
+$(document).on('submit', '#form-search','/search.html' );
+
 
 //function listBeer() {
    // console.log('searching for beer🕵🏿‍♂️...')
