+++ routes/api-routes.js
@@ -1,8 +1,11 @@
 // Requiring our models and passport as we've configured it
 var db = require("../models");
 var passport = require("../config/passport");
-var axios = require("axios");
-var axios = require('post-api-routes')
+
+//importing library
+const Beer = require("../lib/beer.js")
+
+
 module.exports = function (app) {
 
   // Using the passport.authenticate middleware with our local strategy.
@@ -54,50 +57,10 @@ module.exports = function (app) {
   // here we make an axios call to query the database for a beer
   // first we define the url for the api call
 
-  app.get("/api/favorites/search/:id", saveBeer.then())
-
-
-
-
-  function searchBeer(req, res) {
-
-     req = $('#forum-search').val()
-     rank = $('#rank').val()
-
-
-     req = { name: req,
-      rank: rank
-    }
-    
-    const beerData = `https://api.brewerydb.com/v2/beers?key=7873bf684e7db7e59e55ea9dbc1e8d4e&name=${req.name}`;
-    
-    beerData = axios.get(beerData);
-    
-
-    var beer = new class beer {
-      constructor(Id, rank, name,
-      description, abv, labels)
-
-      {
-
-      Id = beerData.Id  
-      rank = req.rank;
-      name = beerData.name;
-      description = beerData.description
-      abv = beerData.abv
-      labels= beerData.labels;
-
-        };
-      }*/ 
-    }).catch(function (err) {
-      console.log(err);
-      res.status(401).json(err);
-
-    }); 
-    
+  app.get("/api/favorites/search/:id",function() {Beer.saveBeer})  
     // return //renderFavorites(beer)
 
-  };
+  
   // route for getting from favorites based off of id
 
   app.get("/api/favorites/:id", function(req, res) {db.Favorite.findOne({
@@ -124,5 +87,7 @@ app.delete("/api/favorites/:id", function(req, res) {
     console.log("Beer Deleted");
   });
 });
+
+};
